server/requirements.txt
alembic==1.16.2
aniso8601==10.0.1
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-cors==6.0.1
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-dotenv==1.1.1
pytz==2024.2
setuptools==70.3.0
six==1.17.0
SQLAlchemy==2.0.29
sqlalchemy-serializer==1.4.22
typing_extensions==4.14.0
Werkzeug==3.1.3
